fixed 5,6,3

+++ 0x08-recursion/6-is_prime_number.c
@@ -14,7 +14,7 @@ int is_prime_number(int n)
 	
 	unsigned int = i;
 
-	if ((float)(n % 2) != 0)
+	if ((n % 2.0) != 0)
 	{
 		return (1);
